UI design when launching game
This includes the ui for: launch screen, screen for guest login to game, screen for host start game. 
Does not include: actual game page with board and everything, waiting page for guest, waiting page for host. Changes to the Host class will need to be done before this can be implemented. Will do so next.